% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMEP.R
\name{RMEP}
\alias{RMEP}
\title{Calculate energy fluxes and actual evapotranspiration}
\usage{
RMEP(Rn, RnL, qs, Ts, type)
}
\arguments{
\item{Rn}{net radiation(unit:W/m2)}

\item{RnL}{net long-wave radiation(unit:W/m2)}

\item{qs}{specific humidity(unit:hpa)}

\item{Ts}{surface temperature(unit:Celsius)}
}
\value{
A list includes latent heat flux(EMEP),sensible heat flux(HMEP),ground heat flux(GMEP) and evapotranspiration(ETMEP)
}
\description{
Calculate energy fluxes and evapotranspiration based on the maximum entropy production(MEP) model
}
