% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMEP.R
\name{RMEPPET}
\alias{RMEPPET}
\title{Calculate potential ET}
\usage{
RMEPPET(Rn, RnL, Ts, type)
}
\arguments{
\item{Rn}{net radiation(unit:W/m2)}

\item{RnL}{net long-wave radiation(unit:W/m2)}

\item{Ts}{surface temperature(unit:Celsius)}
}
\value{
A list concludes latent heat flux(EMEP),sensible heat flux(HMEP),ground heat flux(GMEP) and potential ET(PETMEP)
}
\description{
Calculate energy fluxes and potential evapotranspiration based on the MEP model
}
