% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shum.R
\name{Shum}
\alias{Shum}
\title{Calculate specific humidity for MEP model input}
\usage{
Shum(TA, PA, RH)
}
\arguments{
\item{TA}{air temperature(unit: deg C)}

\item{PA}{Atmospheric pressure(unit: kPa)}

\item{RH}{Relative humidity(unit: %)}
}
\value{
Specific humidity(unit:kg/kg)
}
\description{
Calculate specific humidity for MEP model input
}
