% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SShum.R
\name{SShum}
\alias{SShum}
\title{Calculate saturated specific humidity for MEP model input}
\usage{
SShum(Ts)
}
\arguments{
\item{Ts}{surface temperature(unit: deg C)}
}
\value{
Saturated specific humidity(unit:kg/kg)
}
\description{
Calculate saturated specific humidity for MEP model input
}
\examples{
SShum(Ts=20)
}
